Adds SBT projects
`target` is where most build artifacts live but if the SBT "project" has
build artifacts too – code that modifies the build environment,
generally SBT plugins – then that directory should be cleared as well.

SBT downloads dependencies to a user-level cache that I believe is not
appropriate for this tool to evaluate since the cache is shared across
all projects the user has interacted with.